pyproject.toml
[project]
name = "interaxions"
version = "0.0.0"
description = "A library for agent and environment protocol interactions"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Lei Zhang", email = "jameszhang2880@gmail.com" }]
keywords = ["agent", "environment", "protocol", "interactions", "interaxions"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = ["pyyaml>=6.0", "jinja2>=3.0", "python-dotenv>=1.0.0", "pydantic>=2.0.0"]

[project.urls]
Homepage = "https://github.com/Hambaobao/interaxions"
Repository = "https://github.com/Hambaobao/interaxions"
Issues = "https://github.com/Hambaobao/interaxions/issues"


[project.optional-dependencies]
argo = ["hera>=5.25.1"]
hf = ["datasets>=2.0.0"]
oss = ["ossdata>=0.3.5"]
dev = ["yapf>=0.40.0", "pytest>=9.0.0", "hera>=5.25.1", "datasets>=2.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
